requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
msal==1.26.0
requests==2.31.0
Flask-Session==0.6.0
pandas==2.2.1
xlsxwriter==3.1.9
beautifulsoup4==4.12.3 